requirements.txt

cryptography==46.0.3
dj-rest-auth==7.0.1
django-allauth==65.12.0
django-cors-headers==4.9.0
google-analytics-admin==0.26.0
google-api-python-client==2.184.0
google-auth-oauthlib==1.2.2
grpcio-status==1.75.1
PyJWT==2.10.1
python-dotenv==1.1.1
whitenoise==6.11.0
djangorestframework==3.16.1
django-cors-headers==4.9.0